# Just need to check the last byte.