Refer: #71  #75 

old:
```
go test -bench 'FromString*' -benchmem
goos: darwin
goarch: arm64
pkg: github.com/rs/xid
BenchmarkFromString-10    	445015340	         2.618 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/rs/xid	1.807s
```

new:
```
go test -bench 'FromString*' -benchmem
goos: darwin
goarch: arm64
pkg: github.com/rs/xid
BenchmarkFromString-10    	521384266	         2.304 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/rs/xid	1.556s
```